ddl, statistics: support drop table and schema.
