+++ statistics/ddl.go
@@ -17,6 +17,7 @@ import (
 	"fmt"
 
 	"github.com/juju/errors"
+	"github.com/ngaut/log"
 	"github.com/pingcap/tidb/ddl"
 	"github.com/pingcap/tidb/model"
 	"github.com/pingcap/tidb/mysql"
@@ -29,12 +30,16 @@ func (h *Handle) HandleDDLEvent(t *ddl.Event) error {
 	switch t.Tp {
 	case model.ActionCreateTable:
 		return h.insertTableStats2KV(t.TableInfo)
+	case model.ActionDropTable:
+		return h.deleteTableStats2KV(t.TableInfo)
 	case model.ActionAddColumn:
 		return h.insertColStats2KV(t.TableInfo.ID, t.ColumnInfo)
 	case model.ActionDropColumn:
 		return h.deleteHistStatsFromKV(t.TableInfo.ID, t.ColumnInfo.ID, 0)
 	case model.ActionDropIndex:
 		return h.deleteHistStatsFromKV(t.TableInfo.ID, t.IndexInfo.ID, 1)
+	default:
+		log.Warnf("Unsupported ddl event for statistic %s", t)
 	}
 	return nil
 }
@@ -67,6 +72,29 @@ func (h *Handle) insertTableStats2KV(info *model.TableInfo) error {
 		if err != nil {
 			return errors.Trace(err)
 		}
+	}
+	_, err = exec.Execute("commit")
+	return errors.Trace(err)
+}
+
+func (h *Handle) deleteTableStats2KV(info *model.TableInfo) error {
+	exec := h.ctx.(sqlexec.SQLExecutor)
+	_, err := exec.Execute("begin")
+	if err != nil {
+		return errors.Trace(err)
+	}
+	// First of all, we update the version.
+	_, err = exec.Execute(fmt.Sprintf("update mysql.stats_meta set version = %d where table_id = %d ", h.ctx.Txn().StartTS(), info.ID))
+	if err != nil {
+		return errors.Trace(err)
+	}
+	_, err = exec.Execute(fmt.Sprintf("delete from mysql.stats_histograms where table_id = %d", info.ID))
+	if err != nil {
+		return errors.Trace(err)
+	}
+	_, err = exec.Execute(fmt.Sprintf("delete from mysql.stats_buckets where table_id = %d", info.ID))
+	if err != nil {
+		return errors.Trace(err)
 	}
 	_, err = exec.Execute("commit")
 	return errors.Trace(err)
